// src/api/generated/types.ts
export type AL_BaseAnime = {
    id: number
    idMal?: number
    siteUrl?: string
    status?: AL_MediaStatus
    season?: AL_MediaSeason
    type?: AL_MediaType
    format?: AL_MediaFormat
    seasonYear?: number
    bannerImage?: string
    episodes?: number
    synonyms?: Array<string>
    isAdult?: boolean
    countryOfOrigin?: string
    meanScore?: number
    description?: string
    genres?: Array<string>
    duration?: number
    trailer?: AL_BaseAnime_Trailer
    title?: AL_BaseAnime_Title
    coverImage?: AL_BaseAnime_CoverImage
    startDate?: AL_BaseAnime_StartDate
    endDate?: AL_BaseAnime_EndDate
    nextAiringEpisode?: AL_BaseAnime_NextAiringEpisode
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_CoverImage = {
    color?: string
    extraLarge?: string
    large?: string
    medium?: string
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_EndDate = {
    day?: number
    month?: number
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_NextAiringEpisode = {
    airingAt: number
    episode: number
    timeUntilAiring: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_StartDate = {
    day?: number
    month?: number
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_Title = {
    english?: string
    native?: string
    romaji?: string
    userPreferred?: string
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_Trailer = {
    id?: string
    site?: string
    thumbnail?: string
}
// ...
export type AL_MediaFormat = "TV" |
    "TV_SHORT" |
    "MOVIE" |
    "SPECIAL" |
    "OVA" |
    "ONA" |
    "MUSIC" |
    "MANGA" |
    "NOVEL" |
    "ONE_SHOT"
// ...
export type AL_MediaSeason = "WINTER" | "SPRING" | "SUMMER" | "FALL"
// ...
export type AL_MediaStatus = "FINISHED" | "RELEASING" | "NOT_YET_RELEASED" | "CANCELLED" | "HIATUS"

/**
 * - Filepath: internal/api/anilist/models_gen.go
 * - Filename: models_gen.go
 * - Package: anilist
 * @description
 *  Media type enum, anime or manga.
 */
export type AL_MediaType = "ANIME" | "MANGA"